GIT_SILENT: Increase version number
Increasing the version number is important, otherwise the scripts
that run on kate-editor.org t provide the updates in the download
dialog in Kate's config settings will not pick this update up.

Besides that, we want each released version that has changes to
be distinct, otherwise it will be hard for us to track bugs and/
or regressions... So please always remember to increase the
version number - Thanks! :-)

CCMAIL: guoyunhebrave@gmail.com