Theme: Remove license and author fields
Background: The licensing is a mess for the highlighting files,
since we have many different licenses that may even be incompatible
among each other. To not repeat this error with the Themes, let's
simply remove this field.

The author field is also problematic, since "The Kate Authors"
does not tell anything at all. Better use git to find out the
author, if really necessary.

Differential Revision: https://phabricator.kde.org/D6913