Rewrite the Sieve syntax definition
Summary:
This brings the highlighting up-to-date with RFC 5228 as well as all
currently published extensions, and adds a few features found in the
handwritten highlighter in KDE PIM. It also adds support for variables
and encoded characters in strings, and fixes folding of multi-line
dot-stuffed strings.

A comprehensive test case based on examples from all relevant RFCs is
also included.

This now outperforms the handwritten highlighter in KDE PIM, and looks
also decent on dark color schemes, so we can hopefully port the PIM Sieve
editor to this :)

Reviewers: #framework_syntax_hightlighting, mlaurent, dhaumann

Reviewed By: #framework_syntax_hightlighting, dhaumann

Subscribers: dhaumann, #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D7656