Add syntax highlighting for AppArmor Profiles.
Summary:
AppArmor is a Linux security module that provides mandatory access control on Linux.
The AppArmor Profiles are plain text files that contain access rules for the applications monitored by AppArmor. These are located in: /etc/apparmor.d/

More information:
 * http://wiki.apparmor.net
 * http://manpages.ubuntu.com/manpages/zesty/man5/apparmor.d.5.html

This syntax highlighting file has the characteristic that has a different context for each type of AppArmor rule, so as to avoid conflicts between keywords with different uses and to provide a specific highlighting in each rule.

The highlighting file is made for the latest version of AppArmor: 2.11. Note that by the end of this year (at the latest), AppArmor 4.0 will be released, so, at that time, I will need to update this syntax file with the new features.

Test Plan: This file was created in late February 2017 for personal use and I have been using it to create/edit AppArmor profiles during this time. Therefore, I can assure you that the highlighting file has been tested enough.

Reviewers: #kate, #framework_syntax_highlighting, dhaumann

Reviewed By: #kate, #framework_syntax_highlighting, dhaumann

Subscribers: dhaumann, cullmann, #framework_syntax_highlighting, #frameworks

Tags: #frameworks, #framework_syntax_highlighting

Differential Revision: https://phabricator.kde.org/D7461