JavaScript: include standard mime types
Add application/javascript and text/javascript mimetypes to
javascript.xml

application/javascript is the main one used in shared-mime-info,
without it syntax highlighting is not working for *.mjs files.

text/javascript is added as a safeguard in case if shared-mime-info
switches to it later, as there is ambiguity between which of those two
should be the main one, and IANA could switch to text/javascript.

Refs:
 * https://cgit.freedesktop.org/xdg/shared-mime-info/tree/freedesktop.org.xml.in
 * https://www.iana.org/assignments/media-types/application/javascript
 * https://tools.ietf.org/html/rfc4329
 * https://html.spec.whatwg.org/multipage/scripting.html#javascript-mime-type
 * https://datatracker.ietf.org/doc/draft-ietf-dispatch-javascript-mjs

Reviewers: vkrause

Reviewed By: vkrause

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D8451