Remove unused contexts and itemDatas
Summary:
Some contexts are commented out since they are used in commented
out rules. Besides that, contexts that are not referenced at all
(i.e. also not in other comments), are simply removed entirely.

In addition, searching for lineEmptyContext was missing in the
highlighting indexer - this is added now.

Test Plan: make && make test

Reviewers: vkrause

Reviewed By: vkrause

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D9186