Highlighting indexer: Warn about context switch fallthroughContext="#stay"
Summary:
A fallthroughContext="#stay" is always wrong if fallthrough="true" is set,
since in this case we risk an infinite loop. This patch extends the indexer
to check for this mistakes. Luckily, this mistake does not exist in the
definition files shipped with the KSyntaxHighlighting framework.

Test Plan: make test, manual testing with fallthroughContext="#stay"

Reviewers: vkrause, cullmann

Reviewed By: vkrause

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D9493