Add "None" highlighting to repository per default
Rational: e.g. KTextEditor uses the repository to manage all syntax definitions.
But to allow the user to use "None" as highlighting, we special case that everywhere and use there Definition().
I think it would be much cleaner if the "None" definition is just part of the repository.
An other approach would be to be able to add it later ourself.

Differential Revision: https://phabricator.kde.org/D14762