Add syntax-highlighting for RDoc (R documentation)
Summary:
The R documentation language follows LaTeX-like syntax with occasional
snippets of R code. Builtin keywords and user-defined macros both begin
with a backlash, but we try to distinguish them to highlight arguments
according to their type.

For macros we can't know the argument types (LaTeX/R-like or verbatim),
so we just assume the type is the same as that of the surrounding code.

Reviewers: dhaumann, cullmann

Reviewed By: cullmann

Subscribers: #framework_syntax_highlighting, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D14639