Bash: fix parameter & brace expansion
Summary:
BUG: 387915

Sometimes, the content in braces `{ ... }` is not correctly highlighted, since the RegExpr rule that detects it is very general.
{F6246660}

Also, some parameter expressions are added, such as `${parameter^^pattern}`, `${parameter,,pattern}` & `${parameter@operator}` (before, that was highlighted as "Error"). In the expressions `${parameter:offset:length}`, spaces are allowed after ":" and +/- signs in numbers.

**Source:**
* Bash Reference Manual: https://www.gnu.org/software/bash/manual/bashref.html#Brace-Expansion

Reviewers: cullmann, dhaumann, #framework_syntax_highlighting

Reviewed By: cullmann, dhaumann, #framework_syntax_highlighting

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D15390