AppArmor: fix crashes in open rules (in KF5.50) and improvements for the new Solarized schemes
Summary:
After upgrading to KDE Frameworks 5.50.0, Kate crashes when detecting open AppArmor rules (the AppArmor rules are closed with a comma. If the user doesn't write a comma, the syntax highlighting underlines the statement of the following rule. This is defined in the `_end_rule_irnc` context of the XML file.)

The problem occurs with the use of `#pop#pop!context` in the `_end_rule_irnc` context, only in KF5.50. However, apparently, this only happens in very large files. For example, this problem occurs to me when I open the new test file, updated in this diff. However, I do not know the exact cause of this error...

On line 184 of the test file, I put some examples of open AppArmor rules.

Also, some changes are made in the colors so that the highlighting looks better in the last update of Solarized schemes.

Reviewers: #framework_syntax_highlighting, dhaumann, cullmann

Reviewed By: #framework_syntax_highlighting, cullmann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D15773