keyword rule: Spport for keywords inclusion from another language/file
Summary:
The goal is to reuse words from other definition files without having to copy them. For example, the list of CSS properties also in SCSS.

The presence of `##` in the name of the keyword rule indicates that it is necessary to look for the list in another file (format: `listName##languageName`).

Test Plan: A quick test was done by removing the `properties` list from scss.xml and with `properties##CSS` as the keyword name.

Reviewers: #framework_syntax_highlighting, cullmann, dhaumann, vkrause

Reviewed By: #framework_syntax_highlighting, cullmann

Subscribers: kwrite-devel, vkrause, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D14632