WML: fix embedded Lua code & use new default styles
Summary:
In the previous diff of WML (D17891), I forgot to update the styles as mentioned by dhaumann in the comment (sorry!).

I also noticed that the Lua code is incorrectly highlighted, due to the incorporation of the Shebang in `lua.xml` (`wml.xml` is older). Searching, this is the only file with the inclusion of ##Lua.

Reviewers: #framework_syntax_highlighting, dhaumann

Reviewed By: #framework_syntax_highlighting, dhaumann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D17957