CubeScript: fixes line continuation escape in strings
Summary:
Fix the line continuation escape in strings.

CubeScript files use the extension `.cfg`, which is used by default by Nagios, making it impossible to add a test file. It might be a good idea to add the extension `*.cubescript` and use it for the test file.

Reviewers: #framework_syntax_highlighting, dhaumann, cullmann

Reviewed By: #framework_syntax_highlighting, cullmann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D18948