Logcat: fix dsError color with underline="0"
Summary:
In itemData's, using `dsError` with `underline="0"` doesn't work, the text is underlined.
Therefore, dsError is changed by dsWarning.

I don't know the origin of this bug, this is a temporary solution for the Logcat highlighter.
This bug isn't noticeable in the autotests files, there are no changes when run `make test`.

Test Plan: make test

Reviewers: #framework_syntax_highlighting, dhaumann, cullmann

Reviewed By: #framework_syntax_highlighting, cullmann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D19461