Allow line end context switches in empty lines
Summary:
BUG: 405903

### Bug: Don't skip empty lines after matching a LineContinue rule:

When you write an empty line after a line continuation character, the empty lines are ignored.
This generates an incorrect highlight, since it includes within the context multiple lines after a line continuation character, when the respective context has `lineEndContext="#pop"`.

{F6739814}

The problem is caused because the `lineEndContext` attribute doesn't work in empty lines. Now `lineEndContext` works in empty lines, but only if `lineEmptyContext` is #stay.

Reviewers: #framework_syntax_highlighting, cullmann, dhaumann, vkrause

Reviewed By: #framework_syntax_highlighting, cullmann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D20148