Remove .conf extension from "INI Files" hl, to determine the highlighter using MIME type
Summary:
BUG: 400290

The extension `.conf` is very generic, since depending on the context, it can be used with a different language, and assigning it to the "INI Files" highlighter isn't very convenient.
For example, in "Fontconfig" the XML files have a `.conf` extension.

Therefore, I propose that the `.conf` extension isn't assigned to any syntax highlighter, so use the MIME type to determine the highlighting mode in those files.

For example, when applying this patch, the Fontconfig files (located in `/etc/fonts/`) are highlighted as XML by default.

Reviewers: #framework_syntax_highlighting, dhaumann, cullmann

Reviewed By: #framework_syntax_highlighting, cullmann

Subscribers: kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D21288