Add syntax definition for the Common Intermediate Language
Summary:
Nothing fancy, a lot of details could probably improved, e.g. regarding
the use of '.' as delimiter (which it is for some keywords, while others
start with a '.' themselves). It's however a significant improvement over
no highlighting at all when looking at a lot of Mono disassembler output.

Reviewers: cullmann

Reviewed By: cullmann

Subscribers: cullmann, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D21368