Support for native Matlab strings
Summary:
As of Matlab 2016b, a separate string data type and literal was introduced
in Matlab.

https://de.mathworks.com/help/matlab/matlab_prog/create-string-arrays.html

The new literals are syntactically equivalent to the old character verctors
except that they are delimited by double quote. The term 'character vector'
is now consistently used for the old single quote literales.

The patch renames single quote literals to CharVector and adds double
quote literals as String.

CharVector's are associated with the theme category dsSpecialString.
String's are associated with the theme category dsString. Incomplete
CharVector's and String's are consequently associated with the theme
category dsError.

The examples at the end of the file were extended by the new literals.

Reviewers: cullmann

Reviewed By: cullmann

Subscribers: dhaumann, kwrite-devel, kde-frameworks-devel

Tags: #kate, #frameworks

Differential Revision: https://phabricator.kde.org/D23500