Add POSIX alternate for simple variable assignment
This commit adds '::=' as an operator which defines the value of a
variable to the list of strings which trigger the application of the
Variable attribute in the KSyntaxHighlighting framework. This element of
POSIX syntax is recognized by GNU make v4.0 and later, but not BSD
bmake, however failure to fully adopt an international standard by one
member of a family of related tools should not present an impediment to
widespread adoption elsewhere.

Background:
In 2012, the IEEE published a corrigendum (Latin: a document which
exists solely to correct the errors of an earlier publication) which was
subsequently adopted as [Std 1003.1-2008/Cor 1-2013][1], known better in
computing circles as the document which amended the original POSIX
standard released in 2008 and defined the replacement "POSIX 2013"
specification. Among the changes it made was the addition of '::=' as an
alternate syntax of for the standard '=' variable assignment operator.

Recognition of this operator as a syntactic equivalent for ':=' to
indicate simple (non-recursive) variable assignment in Makefiles was
[added to GNU make][2] in the first release after POSIX 2013 was adopted
by the IEEE: [version 4.0, on 2013-10-09][3].

[1]: https://pubs.opengroup.org/onlinepubs/9699919799.2013edition/
[2]: https://www.gnu.org/software/make/manual/html_node/Flavors.html#Flavors
[3]: https://git.savannah.gnu.org/cgit/make.git/tree/NEWS?h=4.0&id=52191d9d613819a77a321ad6c3ab16e1bc73c381#n65

Signed-off-by: Peter J. Mello <admin@petermello.net>