SPDX comments: fix broken highlighting for multi-line comments
Version 1 of the SPDX comments highlighting rules may break the syntax highlighting
for languages that support multi-line comments.

As there was a recent update to make highlighting files include `Comments` instead
of `Alerts`, all highlighting files supporingt multi-line comments are probably affected
(verified for C++ and XML).

If the mulit-line comment's end-token is on the same line as a SPDX tag,
the SPDX rules will consume the end-token and the highlighting engine stays
in comment mode for the rest of the file.

For `SPDX-License-Identifier` triggering the license-expression context,
this can be fixed by adding fall-through and space detection.

For the remaining SPDX tags there seems to be no way to fix this other than completely
avoiding highlighting the associated value (we would need a generic way to match
the comment's end-token).
The tag itself is still highlighted.