Added support for fluent language files
This aims to add support for projectfluent.org language files.
It's heavily WIP as I'm new to writing syntax highlighting files and
down't know how to do https://www.projectfluent.org/fluent/guide/multiline.html
multiline and block text. the context MultiLineText is intended to be
used for this purpose, but it's currently not working. Any further
comments are appreciated. Furthermore there is currently no
highlighting for selsctors and attributes.