PL/I: Also recognise "pl1" extension and MIME type
The file extension ".pl1" is used for Multics source (along with
".cds" and ".rd" which are also PL/I source, but are rarer and may
clash with other types so are not added).

The original MIME type "text/x-pli" is not defined in shared-mime-info
and does not seem to be used anywhere online.  The type "text/x-pl1src"
has been defined and used for Multics files by
https://github.com/martenjj/multics-mimetypes.

Adding these definitions allows Multics PL/I source files to be
recognised and highlighted.