DefinitionRef: add move constructor and move assignment operator
The user-declared destructor prevented the compiler from generating the
move operations implicitly. Therefore, simply removing the destructor's
declaration and definition makes DefinitionRef movable (and it remains
copyable too).

The rationale and the implicit definition are similar to the case of
Definition detailed in the previous commit message.

Remove a forward declaration of a nonexistent class DefinitionPrivate.