Move the default repository management out of KQuickSyntaxHighlighter
The repository is now a property and can be set externally if needed as
well. This is of limited use at the moment, with the C++ side no being
the same type yet, but with KF6 this will enable sharing a single instance
on both sides.

If no repository is set explicitly, an internally managed default instance
is used, like before.